feat(#80): PR 1 — warm-up dead-alias deletion (net -442 LOC)
Removes the five dead aliases / migration helpers called out in the
streamlining track's PR 1 (issue #80):

- `lore new-wiki <name>` legacy alias (inline `scaffold_wiki` into
  `wiki_cmd.py`; drop `new_wiki_cmd.py`; update README + doctor hints).
- Hidden `lore session new` Typer command; add a no-op `@app.callback`
  so `lore session commit` stays addressable after the dispatcher
  collapses to a single command.
- Hidden `lore hook live-state` alias (use `lore hook context-log`).
- `breadcrumb.migrate_legacy_pending_breadcrumb` + the
  `pending-breadcrumb.txt` legacy file constant; delete its dedicated
  test module and rewrite the staleness test against the JSONL
  hook-events log.
- `synthesis.flush_buffer` re-export; migrate 22 `tests/test_synthesis.py`
  callers + two prose references to `synth_and_close`.

Full test suite green except for two pre-existing failures unrelated
to this PR (`openai` optional dep missing; resume subtree fixture).
