+++ lib/lore_cli/wiki_cmd.py
@@ -3,15 +3,19 @@
 The canonical home for wiki-lifecycle verbs going forward. Today it
 hosts ``new`` (scaffold a new wiki); future work can land alongside
 without inventing more top-level CLI verbs.
-
-The legacy ``lore new-wiki <name>`` form is still accepted as an
-alias and forwards to the same ``scaffold_wiki()`` implementation.
 """
 from __future__ import annotations
 
+import shutil
+import subprocess
+import sys
 from enum import Enum
+from importlib import resources
+from pathlib import Path
 
 import typer
+from lore_core.config import get_wiki_root
+from lore_core.io import atomic_write_text
 from rich.console import Console
 
 from lore_cli._argv_compat import argv_main
@@ -27,11 +31,81 @@ app = typer.Typer(
 )
 
 
-# Re-import the shared mode enum so callers don't have to thread it
-# through both modules.
 class WikiMode(str, Enum):
     personal = "personal"
     team = "team"
+
+
+SUBDIRS = ("projects", "concepts", "decisions", "sessions", "inbox")
+TEMPLATE_NAMES = ("standard", "science", "design", "custom")
+
+
+def _plugin_templates_dir() -> Path:
+    from lore_core.templates import templates_dir
+    return templates_dir()
+
+
+def _load_template(name: str) -> str:
+    if name not in TEMPLATE_NAMES:
+        raise ValueError(f"unknown template {name!r}; choose from {TEMPLATE_NAMES}")
+    return resources.files("lore_core.surface_templates").joinpath(f"{name}.md").read_text()
+
+
+def scaffold_wiki(
+    name: str,
+    *,
+    mode: str = "personal",
+    remote: str | None = None,
+    force: bool = False,
+    surfaces: str = "standard",
+) -> Path:
+    wiki_root = get_wiki_root()
+    wiki_root.mkdir(parents=True, exist_ok=True)
+    target = wiki_root / name
+
+    is_new_wiki = not target.exists()
+
+    if is_new_wiki or force:
+        target.mkdir(exist_ok=True)
+        for sub in SUBDIRS:
+            (target / sub).mkdir(exist_ok=True)
+
+        templates_src = _plugin_templates_dir()
+        claude_md = target / "CLAUDE.md"
+        claude_md.write_text((templates_src / "wiki-CLAUDE.md").read_text())
+        (target / "templates").mkdir(exist_ok=True)
+        shutil.copy(templates_src / "session.md", target / "templates" / "session.md")
+
+        (target / "_index.txt").write_text(
+            f"# {name.upper()} Knowledge Index\n\n"
+            f"(Newly created wiki — run `lore lint --wiki {name}` to populate.)\n"
+        )
+
+    surfaces_path = target / "SURFACES.md"
+    if not surfaces_path.exists():
+        if surfaces not in TEMPLATE_NAMES:
+            err_console.print(f"[red]unknown template '{surfaces}'; choose from {TEMPLATE_NAMES}[/red]")
+            raise typer.Exit(1)
+        atomic_write_text(surfaces_path, _load_template(surfaces))
+
+    if mode == "team":
+        subprocess.run(["git", "init"], cwd=str(target), check=False)
+        if remote:
+            subprocess.run(
+                ["git", "remote", "add", "origin", remote],
+                cwd=str(target),
+                check=False,
+            )
+        subprocess.run(["git", "add", "-A"], cwd=str(target), check=False)
+        subprocess.run(
+            ["git", "commit", "-m", "lore: initial wiki scaffold"],
+            cwd=str(target),
+            check=False,
+        )
+
+    console.print(f"[green]Created {target}[/green]")
+    console.print(f"Next: run [cyan]lore lint --wiki {name}[/cyan] to regenerate catalogs.")
+    return target
 
 
 @app.command("new")
@@ -51,14 +125,15 @@ def cmd_new(
     surfaces: str = typer.Option(
         "standard",
         "--surfaces",
-        help="SURFACES.md template: standard | science | design | custom",
+        help=f"SURFACES.md template: {TEMPLATE_NAMES}",
     ),
 ) -> None:
     """Scaffold a new wiki under $LORE_ROOT/wiki/."""
-    # Lazy import to keep this module's import surface narrow.
-    from lore_cli.new_wiki_cmd import scaffold_wiki
-
     scaffold_wiki(name, mode=mode.value, remote=remote, force=force, surfaces=surfaces)
 
 
 main = argv_main(app)
+
+
+if __name__ == "__main__":
+    sys.exit(main())
